requirements.txt
flask>=3.0.0
ultralytics>=8.2.0
opencv-python-headless>=4.9.0
matplotlib>=3.8.0
pandas>=2.2.0
python-dotenv>=1.0.1
requests>=2.31.0
langchain>=0.2.12
langchain-community>=0.2.11
langchain-groq>=0.1.6
langchain-text-splitters>=0.2.2
langchain-ollama>=0.1.1
chromadb>=0.5.3
